# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Slots API",
    "version": "1.0.0",
    "description": "Slots RGS (Remote Game Server) API.\n\n**Auth flow (two-token pattern):**\n1. `POST /api/v1/auth/register` or `/login` — returns a 15-min **access token** in the body; sets a 7-day **refresh token** in an `httpOnly` cookie\n2. Copy `access_token` → click **Authorize** → paste → game endpoints unlock\n3. When the access token expires, `POST /api/v1/auth/refresh` silently issues a new pair (token rotation)\n4. `POST /api/v1/auth/logout` revokes all refresh tokens and clears the cookie"
  },
  "servers": [
    {
      "url": "https://pyavchik.space",
      "description": "Production"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "additionalProperties": false
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "player@example.com",
            "description": "Player email address"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "password123",
            "description": "Min. 8 characters"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "player@example.com",
            "description": "Player email address"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "RS256 JWT — paste into the Authorize dialog above"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ],
            "example": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "example": 900,
            "description": "Access token TTL in seconds (15 min)"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ],
        "additionalProperties": false
      },
...
      "RouletteSpinRequest": {
        "type": "object",
        "properties": {
          "session_id": {
            "type": "string"
          },
          "game_id": {
            "type": "string",
            "default": "roulette_european_001"
          },
          "bets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "enum": [
                    "straight",
                    "split",
                    "street",
                    "trio",
                    "corner",
                    "basket",
                    "sixLine",
                    "column",
                    "dozen",
                    "red",
                    "black",
                    "even",
                    "odd",
                    "high",
                    "low"
                  ]
                },
                "numbers": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 36
                  }
                },
                "amount": {
                  "type": "number",
                  "minimum": 0,
                  "exclusiveMinimum": true
                }
              },
              "required": [
                "type",
                "numbers",
                "amount"
              ],
              "additionalProperties": false
            },
            "minItems": 1,
            "maxItems": 200
          },
          "client_timestamp": {
            "type": "number"
          }
        },
        "required": [
          "session_id",
          "bets"
        ],
        "additionalProperties": false
      },
      "RouletteSpinResponse": {
        "type": "object",
        "properties": {
          "spin_id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "game_id": {
            "type": "string"
          },
          "balance": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "number"
              },
              "currency": {
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ]
          },
          "outcome": {
            "type": "object",
            "properties": {
              "winning_number": {
                "type": "integer",
                "minimum": 0,
                "maximum": 36
              },
              "winning_color": {
                "type": "string",
                "enum": [
                  "red",
                  "black",
                  "green"
                ]
              },
              "wheel_position": {
                "type": "integer"
              },
              "win": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "currency": {
                    "type": "string"
                  },
                  "breakdown": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "bet_type": {
                          "type": "string",
                          "enum": [
                            "straight",
                            "split",
                            "street",
                            "trio",
                            "corner",
                            "basket",
                            "sixLine",
                            "column",
                            "dozen",
                            "red",
                            "black",
                            "even",
                            "odd",
                            "high",
                            "low"
                          ]
                        },
                        "numbers": {
                          "type": "array",
                          "items": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 36
                          }
                        },
                        "bet_amount": {
                          "type": "number"
                        },
                        "payout": {
                          "type": "number"
                        },
                        "profit": {
                          "type": "number"
                        },
                        "la_partage": {
                          "type": "boolean"
                        },
                        "won": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "bet_type",
                        "numbers",
                        "bet_amount",
                        "payout",
                        "profit",
                        "la_partage",
                        "won"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "required": [
                  "amount",
                  "currency",
                  "breakdown"
                ]
              },
              "total_bet": {
                "type": "number"
              },
              "total_return": {
                "type": "number"
              }
            },
            "required": [
              "winning_number",
              "winning_color",
              "wheel_position",
              "win",
              "total_bet",
              "total_return"
            ],
            "additionalProperties": false
          },
          "timestamp": {
            "type": "number"
          }
        },
        "required": [
          "spin_id",
          "session_id",
          "game_id",
          "balance",
          "outcome",
          "timestamp"
        ],
        "additionalProperties": false
      },
      "RouletteInitResponse": {
        "type": "object",
        "properties": {
          "session_id": {
            "type": "string"
          },
          "game_id": {
            "type": "string",
            "default": "roulette_european_001"
          },
          "config": {
            "type": "object",
            "properties": {
              "game_id": {
                "type": "string",
                "default": "roulette_european_001"
              },
              "type": {
                "type": "string",
                "enum": [
                  "roulette"
                ]
              },
              "variant": {
                "type": "string",
                "enum": [
                  "european"
                ]
              },
              "numbers": {
                "type": "integer",
                "default": 37
              },
              "min_bet": {
                "type": "number"
              },
              "max_total_bet": {
                "type": "number"
              },
              "bet_levels": {
                "type": "array",
                "items": {
                  "type": "number"
                }
              },
              "bet_types": {
                "type": "object",
                "additionalProperties": {
                  "type": "object",
                  "properties": {
                    "payout": {
                      "type": "number"
                    },
                    "size": {
                      "type": "number"
                    },
                    "maxBet": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "payout",
                    "size",
                    "maxBet"
                  ]
                }
              },
              "currencies": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "rtp": {
                "type": "number"
              },
              "features": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "wheel_order": {
                "type": "array",
                "items": {
                  "type": "integer"
                }
              },
              "number_colors": {
                "type": "object",
                "additionalProperties": {
                  "type": "string",
                  "enum": [
                    "red",
                    "black",
                    "green"
                  ]
                }
              }
            },
            "required": [
              "type",
              "variant",
              "min_bet",
              "max_total_bet",
              "bet_levels",
              "bet_types",
              "currencies",
              "rtp",
              "features",
              "wheel_order",
              "number_colors"
            ],
            "additionalProperties": false
          },
          "balance": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "number"
              },
              "currency": {
                "type": "string"
              }
            },
            "required": [
              "amount",
              "currency"
            ]
          },
          "recent_numbers": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "expires_at": {
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "config",
          "balance",
          "recent_numbers",
          "expires_at"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new player account",
        "description": "Creates an account and issues a token pair:\n- **Body**: 15-min RS256 access token — copy and paste into **Authorize**\n- **Cookie** (`httpOnly`): 7-day refresh token — stored by the browser, never readable by JS",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created, JWT issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login and receive a JWT",
        "description": "Verifies credentials and issues a fresh token pair. Same response shape as register.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Silently refresh the access token",
        "description": "Reads the `refresh_token` httpOnly cookie, validates it, and issues a new token pair (**token rotation** — the old refresh token is invalidated on use). Returns 401 if the cookie is missing or expired. In the browser this happens automatically; the cookie is never visible to JS.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "refresh_token=<opaque-hex>",
              "description": "Set automatically by the browser"
            },
            "required": false,
            "description": "Set automatically by the browser",
            "name": "Cookie",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "New access token issued, new refresh cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/roulette/init": {
      "post": {
        "tags": [
          "Roulette"
        ],
        "summary": "Initialize a roulette session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouletteInitResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roulette/spin": {
      "post": {
        "tags": [
          "Roulette"
        ],
        "summary": "Place roulette bets and spin the wheel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouletteSpinRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spin resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouletteSpinResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or balance error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}